requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15
django-filter>=24.3
drf-spectacular>=0.27
django-import-export>=4.1
django-jazzmin>=3.0
psycopg[binary]>=3.2
pandas>=2.2
openpyxl>=3.1
celery>=5.4
redis>=5.0
Pillow>=10.4
gunicorn>=23.0
python-dotenv>=1.0
dj-database-url>=2.2
whitenoise>=6.7
